At this point - the pre-req checking isn't working very well
Problem is - when I hit a pre-req that's required - I stick up the modal
and ask for a decision - but the checking loops on to check the next
pre-req and if that also needs to prompt - well the modal is in use.
So I somehow need to check to see how many  missing pre-reqs we have -
if more than 1 take the first one and prompt and finish. Then repeat
that process till there are NO missing pre-reqs